Qayta qayta yozmaslik uchun
Aslida esa Generatsiya qilish ham shart emas
Birinchi existligini tekshirish kerak

+++ src/EntityCore.Tools/Manager.cs
@@ -192,14 +192,21 @@ namespace EntityCore.Tools
 
             Directory.CreateDirectory(directoryPath);
             string filePath = Path.Combine(directoryPath, fileName);
+
+            if (File.Exists(filePath))
+            {
+                ConsoleMessage($"!!! {fileName} already exists.", ConsoleColor.Yellow);
+                return;
+            }
+
             File.WriteAllText(filePath, code);
 
             ConsoleMessage($"{fileName} generated successfully!");
         }
 
-        private void ConsoleMessage(string message)
+        private void ConsoleMessage(string message, ConsoleColor consoleColor = ConsoleColor.Green)
         {
-            Console.ForegroundColor = ConsoleColor.Green;
+            Console.ForegroundColor = consoleColor;
             Console.WriteLine(message);
             Console.ResetColor();
         }
